fix(sdcm/utils/docker.py): Rename to docker_utils.py to avoid name confusion with docker package
(cherry picked from commit bebff93f29e38fdf8359138bd1e7f0545c023883)
