+++ sdcm/cdclog_reader_thread.py
@@ -8,7 +8,7 @@ from typing import List, Dict
 from sdcm.sct_events import CDCReaderStressEvent, Severity
 from sdcm.utils.common import get_docker_stress_image_name
 from sdcm.utils.thread import DockerBasedStressThread
-from sdcm.utils.docker import RemoteDocker
+from sdcm.utils.docker_utils import RemoteDocker
 from sdcm.stress_thread import format_stress_cmd_error
 
 LOGGER = logging.getLogger(__name__)
