+++ unit_tests/conftest.py
@@ -6,7 +6,7 @@ from pathlib import Path
 import pytest
 
 from sdcm.prometheus import start_metrics_server
-from sdcm.utils.docker import RemoteDocker
+from sdcm.utils.docker_utils import RemoteDocker
 from sdcm import wait
 from sdcm.utils.decorators import timeout
 from sdcm.sct_events import (start_events_device, stop_events_device)
